LayoutTests/canvas/philip/tests/2d.gradient.radial.nonfinite-expected.txt

< [index] >
2d.gradient.radial.nonfinite
createRadialGradient() throws NOT_SUPPORTED_ERR if arguments are not finite
References: 2d.gradient.radial.nonfinite
Actual output:
Passed
